Wraps command-line arguments parsing with a function (#235)
The arguments are parsed automagically because the parser and the act of
parsing is executed during the import time, when models.PyCryptoBot is
imported. This becomes a problem when attempting imports from tests
running undr pytest: pytest arguments are what parser sees in `sys.argv`,
and it's not possible to *not* parse them which leads to an error,
usually parser reporting "unrecognized arguments".

Therefore wrapping the parser & co. with a function, called when
arguments are needed.