Docker Option 2 Update - Code and File Structure
I noticed when I was trying to use Docker for multiple markets, it wasn't working. It kept trying to read the coinbase.key file but couldn't. after a little tickering, 

i got it to work by updating the docker-compose.yaml to include the coinbase.key under volumes. I think Docker wasn't able to read the file because it wasn't being pulled in when being built/composed.

Take a look at the changes I made to this file and confirm it's the correct way of fixing this. I also added verbiage regarding adding the coinbase.key file to each folder. not sure if you want it like that but thought I would add it.

Just started using this a couple of weeks ago and love it. Thank you for all your work and I will contribute as much as i can.